feat: update model with int64 ID and complete gorm tags
- BaseModel: change ID from uint to int64
- All foreign keys: change from uint to int64
- Add missing gorm column tags to all fields
- Add CIConfig, CDConfig, CanaryConfig models
- Update InstanceConfig with InstanceSpec wrapper for K8s native specs

+++ backend/infra/mysql/model/dependency_binding.go
@@ -3,8 +3,8 @@ package model
 // DependencyBinding 依赖绑定 - 实例配置与依赖的绑定关系
 type DependencyBinding struct {
 	BaseModel
-	InstanceConfigID uint `gorm:"column:instance_config_id;not null;index" json:"instance_config_id"`
-	DependencyID     uint `gorm:"column:dependency_id;not null;index" json:"dependency_id"`
+	InstanceConfigID int64 `gorm:"column:instance_config_id;not null;index" json:"instance_config_id"`
+	DependencyID     int64 `gorm:"column:dependency_id;not null;index" json:"dependency_id"`
 }
 
 func (DependencyBinding) TableName() string {
